address review comments - build in ci
