+++ common/src/test/scala/uk/ac/wellcome/platform/archive/common/fixtures/StorageRandomThings.scala
@@ -26,8 +26,7 @@ trait StorageRandomThings extends RandomThings {
 
   val dummyQueue: Queue = Queue(
     url = "test://test-q",
-    arn = "arn::sqs::test",
-    visibilityTimeout = 1
+    arn = "arn::sqs::test"
   )
 
   def collectionOf[T](min: Int = 0, max: Int = collectionMax)(f: => T): Seq[T] =
