+++ ingests_tracker/src/main/scala/uk/ac/wellcome/platform/storage/ingests_tracker/client/IngestTrackerError.scala
@@ -1,0 +1,11 @@
+package uk.ac.wellcome.platform.storage.ingests_tracker.client
+
+import uk.ac.wellcome.platform.archive.common.ingests.models.IngestUpdate
+
+sealed trait IngestTrackerError {
+  val ingestUpdate: IngestUpdate
+}
+case class IngestTrackerConflictError(ingestUpdate: IngestUpdate)
+  extends IngestTrackerError
+case class IngestTrackerUnknownError(ingestUpdate: IngestUpdate, err: Throwable)
+  extends IngestTrackerError
