+++ ingests_tracker/src/test/scala/uk/ac/wellcome/platform/storage/ingests_tracker/client/IngestTrackerClientTest.scala
@@ -1,4 +1,4 @@
-package client
+package uk.ac.wellcome.platform.storage.ingests_tracker.client
 
 import java.time.Instant
 
@@ -10,11 +10,6 @@ import org.scalatest.matchers.should.Matchers
 import uk.ac.wellcome.platform.archive.common.fixtures.HttpFixtures
 import uk.ac.wellcome.platform.archive.common.generators.IngestGenerators
 import uk.ac.wellcome.platform.archive.common.ingests.models.Ingest.Failed
-import uk.ac.wellcome.platform.storage.ingests_tracker.client.{
-  AkkaIngestTrackerClient,
-  IngestTrackerConflictError,
-  IngestTrackerUnknownError
-}
 import uk.ac.wellcome.platform.archive.common.ingests.models.Ingest.Succeeded
 import uk.ac.wellcome.platform.storage.ingests_tracker.fixtures.IngestsTrackerApiFixture
 
@@ -70,7 +65,7 @@ class IngestTrackerClientTest
     }
   }
 
-  it("conflicts with a valid ingest") {
+  it("fails if the ingest update conflicts with the stored ingest") {
     withIngestsTrackerApi(Seq(failedIngest)) {
       case (_, _, ingestsTracker) =>
         val update = ingestTrackerClient.updateIngest(ingestStatusUpdate)
