+++ ingests_worker/src/test/scala/uk/ac/wellcome/platform/storage/ingests_worker/fixtures/IngestsWorkerFixtures.scala
@@ -66,7 +66,7 @@ trait IngestsWorkerFixtures
       Future.successful(Left(IngestTrackerUnknownError(ingestUpdate, error)))
     )
 
-  def failedFutureClient = new FakeIngestTrackerClient(
+  def failedFutureClient(error: Throwable = error) = new FakeIngestTrackerClient(
     Future.failed(error)
   )
